@@ -1,114 +1,83 @@
 import Foundation
-#if !COCOAPODS
 import ApolloAPI
-#endif
 
-/// A network transport is responsible for sending GraphQL operations to a server.
-public protocol NetworkTransport: AnyObject {
+/// A protocol for a type that is responsible for sending GraphQL query and mutation operations to a server.
+///
+/// To support subscription operations, a ``NetworkTransport`` should also implement the
+/// ``SubscriptionNetworkTransport`` protocol.
+public protocol NetworkTransport: AnyObject, Sendable {
 
-  /// Send a GraphQL operation to a server and return a response.
-  ///
-  /// Note if you're implementing this yourself rather than using one of the batteries-included versions of `NetworkTransport` (which handle this for you): The `clientName` and `clientVersion` should be sent with any URL request which needs headers so your client can be identified by tools meant to see what client is using which request. The `addApolloClientHeaders` method is provided below to do this for you if you're using Apollo Studio.
+  /// Send a GraphQL query to a server and return a response.
   ///
   /// - Parameters:
-  ///   - operation: The operation to send.
-  ///   - cachePolicy: The `CachePolicy` to use making this request.
-  ///   - contextIdentifier:  [optional] A unique identifier for this request, to help with deduping cache hits for watchers. Defaults to `nil`.
-  ///   - context: [optional] A context that is being passed through the request chain. Defaults to `nil`.
-  ///   - callbackQueue: The queue to call back on with the results. Should default to `.main`.
-  ///   - completionHandler: A closure to call when a request completes. On `success` will contain the response received from the server. On `failure` will contain the error which occurred.
-  /// - Returns: An object that can be used to cancel an in progress request.
-  func send<Operation: GraphQLOperation>(operation: Operation,
-                                         cachePolicy: CachePolicy,
-                                         contextIdentifier: UUID?,
-                                         context: (any RequestContext)?,
-                                         callbackQueue: DispatchQueue,
-                                         completionHandler: @escaping (Result<GraphQLResult<Operation.Data>, any Error>) -> Void) -> any Cancellable
+  ///   - query: The `GraphQLQuery` operation to send.
+  ///   - fetchBehavior: The `FetchBehavior` to use for this request.
+  ///   Determines if fetching will include cache/network fetches.
+  ///   - requestConfiguration: A configuration used to configure per-request behaviors for this request
+  /// - Returns: A stream of `GraphQLResult`s for each response.
+  func send<Query: GraphQLQuery>(
+    query: Query,
+    fetchBehavior: FetchBehavior,
+    requestConfiguration: RequestConfiguration
+  ) throws -> AsyncThrowingStream<GraphQLResponse<Query>, any Error>
 
-  /// The name of the client to send as a header value.
-  var clientName: String { get }
+  /// Send a GraphQL mutation to a server and return a response.
+  ///
+  /// - Parameters:
+  ///   - mutation: The `GraphQLMutation` operation to send.
+  ///   - fetchBehavior: The `FetchBehavior` to use for this request.
+  ///   Determines if fetching will include cache/network fetches.
+  ///   - requestConfiguration: A configuration used to configure per-request behaviors for this request
+  /// - Returns: A stream of `GraphQLResult`s for each response.
+  func send<Mutation: GraphQLMutation>(
+    mutation: Mutation,
+    requestConfiguration: RequestConfiguration
+  ) throws -> AsyncThrowingStream<GraphQLResponse<Mutation>, any Error>
 
-  /// The version of the client to send as a header value.
-  var clientVersion: String { get }
-}
-
-public extension NetworkTransport {
-
-  /// The field name for the Apollo Client Name header
-  static var headerFieldNameApolloClientName: String {
-    return "apollographql-client-name"
-  }
-
-  /// The field name for the Apollo Client Version header
-  static var headerFieldNameApolloClientVersion: String {
-    return "apollographql-client-version"
-  }
-
-  /// The default client name to use when setting up the `clientName` property
-  static var defaultClientName: String {
-    guard let identifier = Bundle.main.bundleIdentifier else {
-      return "apollo-ios-client"
-    }
-
-    return "\(identifier)-apollo-ios"
-  }
-
-  var clientName: String {
-    return Self.defaultClientName
-  }
-
-  /// The default client version to use when setting up the `clientVersion` property.
-  static var defaultClientVersion: String {
-    var version = String()
-    if let shortVersion = Bundle.main.shortVersion {
-      version.append(shortVersion)
-    }
-
-    if let buildNumber = Bundle.main.buildNumber {
-      if version.isEmpty {
-        version.append(buildNumber)
-      } else {
-        version.append("+\(buildNumber)")
-      }
-    }
-
-    if version.isEmpty {
-      version = "(unknown)"
-    }
-
-    return version
-  }
-
-  var clientVersion: String {
-    return Self.defaultClientVersion
-  }
-
-  /// Adds the Apollo client headers for this instance of `NetworkTransport` to the given request
-  /// - Parameter request: A mutable URLRequest to add the headers to.
-  func addApolloClientHeaders(to request: inout URLRequest) {
-    request.setValue(self.clientName, forHTTPHeaderField: Self.headerFieldNameApolloClientName)
-    request.setValue(self.clientVersion, forHTTPHeaderField: Self.headerFieldNameApolloClientVersion)
-  }
 }
 
 // MARK: -
 
-/// A network transport which can also handle uploads of files.
-public protocol UploadingNetworkTransport: NetworkTransport {
+/// A protocol for a type that is responsible for sending GraphQL subscriptions to a server.
+///
+/// To support query and mutation operations, a ``SubscriptionNetworkTransport`` should also implement the
+/// ``NetworkTransport`` protocol.
+public protocol SubscriptionNetworkTransport {
 
-  /// Uploads the given files with the given operation.
+  /// Send a GraphQL subscription to a server and return a response stream.
   ///
   /// - Parameters:
-  ///   - operation: The operation to send
+  ///   - subscription: The `GraphQLSubscription` operation to send.
+  ///   - fetchBehavior: The `FetchBehavior` to use for this request.
+  ///   Determines if fetching will include cache/network fetches.
+  ///   - requestConfiguration: A configuration used to configure per-request behaviors for this request
+  /// - Returns: A stream of `GraphQLResult`s for each response.
+  func send<Subscription: GraphQLSubscription>(
+    subscription: Subscription,
+    fetchBehavior: FetchBehavior,
+    requestConfiguration: RequestConfiguration
+  ) throws -> AsyncThrowingStream<GraphQLResponse<Subscription>, any Error>
+
+}
+
+// MARK: -
+
+/// A protocol for a type that is responsible for sending GraphQL file upload operations to a server.
+///
+/// To support query and mutation operations without file uploads, an ``UploadingNetworkTransport`` should also
+/// implement the ``NetworkTransport`` protocol.
+public protocol UploadingNetworkTransport {
+
+  /// Sends a GraphQL operation to a server and uploads the given files.
+  ///
+  /// - Parameters:
+  ///   - operation: The GraphQL operation to send
   ///   - files: An array of `GraphQLFile` objects to send.
-  ///   - context: [optional] A context that is being passed through the request chain.
-  ///   - callbackQueue: The queue to call back on with the results. Should default to `.main`.
-  ///   - completionHandler: The completion handler to execute when the request completes or errors
-  /// - Returns: An object that can be used to cancel an in progress request.
+  ///   - requestConfiguration: A configuration used to configure per-request behaviors for this request
+  /// - Returns: A stream of `GraphQLResult`s for each response.
   func upload<Operation: GraphQLOperation>(
     operation: Operation,
     files: [GraphQLFile],
-    context: (any RequestContext)?,
-    callbackQueue: DispatchQueue,
-    completionHandler: @escaping (Result<GraphQLResult<Operation.Data>,any Error>) -> Void) -> any Cancellable
+    requestConfiguration: RequestConfiguration
+  ) throws -> AsyncThrowingStream<GraphQLResponse<Operation>, any Error>
 }
