@@ -1,7 +1,6 @@
-#if !COCOAPODS
 @_exported import ApolloAPI
-#endif
-@_spi(Execution) import Apollo
+@_spi(Internal) import ApolloAPI
+@_spi(Execution) @_spi(Internal) import Apollo
 import Foundation
 
 @dynamicMemberLookup
@@ -98,12 +97,12 @@ public class Mock<O: MockObject>: AnyMock, Hashable {
   public var _selectionSetMockData: JSONObject {
     _data.mapValues {
       if let mock = $0.base as? (any AnyMock) {
-        return mock._selectionSetMockData
+        return mock._selectionSetMockData as JSONValue
       }
       if let mockArray = $0 as? Array<Any> {
-        return mockArray._unsafelyConvertToSelectionSetData()
+        return mockArray._unsafelyConvertToSelectionSetData() as JSONValue
       }
-      return $0
+      return $0 as JSONValue
     }
   }
 
@@ -124,11 +123,11 @@ public extension RootSelectionSet {
   static func from<O: MockObject>(
     _ mock: Mock<O>,
     withVariables variables: GraphQLOperation.Variables? = nil
-  ) -> Self {
+  ) async -> Self {
     let accumulator = TestMockSelectionSetMapper<Self>()
     let executor = GraphQLExecutor(executionSource: NetworkResponseExecutionSource())
 
-    return try! executor.execute(
+    return try! await executor.execute(
       selectionSet: Self.self,
       on: mock._selectionSetMockData,
       variables: variables,
@@ -144,14 +143,14 @@ public protocol AnyMock {
 }
 
 public protocol MockObject: MockFieldValue {
-  associatedtype MockFields
+  associatedtype MockFields: Sendable
   associatedtype MockValueCollectionType = Array<Mock<Self>>
 
   static var objectType: Object { get }
   static var _mockFields: MockFields { get }
 }
 
-public protocol MockFieldValue {
+public protocol MockFieldValue: Sendable {
   associatedtype MockValueCollectionType: Collection
 }
 
@@ -187,26 +186,27 @@ fileprivate extension Array {
     }
   }
 
-  func _unsafelyConvertToSelectionSetData() -> [AnyHashable?] {
+  func _unsafelyConvertToSelectionSetData() -> [JSONValue?] {
     map(_unsafelyConvertToSelectionSetData(element:))
   }
 
-  private func _unsafelyConvertToSelectionSetData(element: Any) -> AnyHashable? {
+  private func _unsafelyConvertToSelectionSetData(element: Any) -> JSONValue? {
     switch element {
     case let element as any AnyMock:
-      return element._selectionSetMockData
+      return element._selectionSetMockData as JSONValue
 
     case let innerArray as Array<Any>:
-      return innerArray._unsafelyConvertToSelectionSetData()
+      return innerArray._unsafelyConvertToSelectionSetData() as JSONValue
+
+    case let optionalElement as Optional<any Sendable>:
+      guard case let .some(element) = optionalElement.asNullable else {
+        return nil
+      }
+      return _unsafelyConvertToSelectionSetData(element: element)
 
     case let element as AnyHashable:
-      if DataDict._AnyHashableCanBeCoerced {
-        return element
-
-      } else {
-        return _unsafelyConvertToSelectionSetData(element: element.base)
-      }
-
+      return _unsafelyConvertToSelectionSetData(element: element.base)
+      
     default:
       return nil
     }
