@@ -2,8 +2,5 @@ import Foundation
 
 public enum Constants {
   public static let ApolloClientName = "apollo-ios"
-  public static let ApolloClientVersion: String = "1.24.0"
-
-  @available(*, deprecated, renamed: "ApolloClientVersion")
-  public static let ApolloVersion: String = ApolloClientVersion
+  public static let ApolloClientVersion: String = "2.0.0-beta-3"
 }
