@@ -1,7 +1,5 @@
 import Foundation
-#if !COCOAPODS
-import ApolloAPI
-#endif
+@_spi(Internal) @_spi(Unsafe) import ApolloAPI
 
 public enum JSONConverter {
   
@@ -19,7 +17,7 @@ public enum JSONConverter {
   /// Converts a ``GraphQLResult`` into a basic JSON dictionary for use.
   ///
   /// - Returns: A `[String: Any]` JSON dictionary representing the ``GraphQLResult``.
-  public static func convert<T>(_ result: GraphQLResult<T>) -> [String: Any] {
+  public static func convert<T>(_ result: GraphQLResponse<T>) -> [String: Any] {
     result.asJSONDictionary()
   }
 
