@@ -14,9 +14,6 @@ public final class MultipartFormData {
     }
   }
 
-  /// A Carriage Return Line Feed character, which will be seen as a newline on both unix and Windows servers.
-  static let CRLF = "\r\n"
-
   public let boundary: String
 
   private var bodyParts: [BodyPart]
@@ -29,20 +26,17 @@ public final class MultipartFormData {
     self.bodyParts = []
   }
 
-  /// Convenience initializer which uses a pre-defined boundary
-  public convenience init() {
-    self.init(boundary: "apollo-ios.boundary.\(UUID().uuidString)")
-  }
-
   /// Appends the passed-in string as a part of the body.
   ///
   /// - Parameters:
   ///   - string: The string to append
   ///   - name: The name of the part to pass along to the server
   public func appendPart(string: String, name: String) throws {
-    self.appendPart(data: try self.encode(string: string),
-                    name: name,
-                    contentType: nil)
+    self.appendPart(
+      data: try self.encode(string: string),
+      name: name,
+      contentType: nil
+    )
   }
 
   /// Appends the passed-in data as a part of the body.
@@ -52,18 +46,22 @@ public final class MultipartFormData {
   ///   - name: The name of the part to pass along to the server
   ///   - contentType: [optional] The content type of this part. Defaults to nil.
   ///   - filename: [optional] The name of the file for this part. Defaults to nil.
-  public func appendPart(data: Data,
-                         name: String,
-                         contentType: String? = nil,
-                         filename: String? = nil) {
+  public func appendPart(
+    data: Data,
+    name: String,
+    contentType: String? = nil,
+    filename: String? = nil
+  ) {
     let inputStream = InputStream(data: data)
     let contentLength = UInt64(data.count)
 
-    self.appendPart(inputStream: inputStream,
-                    contentLength: contentLength,
-                    name: name,
-                    contentType: contentType,
-                    filename: filename)
+    self.appendPart(
+      inputStream: inputStream,
+      contentLength: contentLength,
+      name: name,
+      contentType: contentType,
+      filename: filename
+    )
   }
 
   /// Appends the passed-in input stream as a part of the body.
@@ -74,16 +72,22 @@ public final class MultipartFormData {
   ///   - name: The name of the part to pass along to the server
   ///   - contentType: [optional] The content type of this part. Defaults to nil.
   ///   - filename: [optional] The name of the file for this part. Defaults to nil.
-  public func appendPart(inputStream: InputStream,
-                         contentLength: UInt64,
-                         name: String,
-                         contentType: String? = nil,
-                         filename: String? = nil) {
-    self.bodyParts.append(BodyPart(name: name,
-                                   inputStream: inputStream,
-                                   contentLength: contentLength,
-                                   contentType: contentType,
-                                   filename: filename))
+  public func appendPart(
+    inputStream: InputStream,
+    contentLength: UInt64,
+    name: String,
+    contentType: String? = nil,
+    filename: String? = nil
+  ) {
+    self.bodyParts.append(
+      BodyPart(
+        name: name,
+        inputStream: inputStream,
+        contentLength: contentLength,
+        contentType: contentType,
+        filename: filename
+      )
+    )
   }
 
   /// Encodes everything into the final form data to send to a server.
@@ -104,10 +108,11 @@ public final class MultipartFormData {
   private func encode(bodyPart: BodyPart) throws -> Data {
     var encoded = Data()
 
-    encoded.append(try self.encode(string: "--\(self.boundary)\(MultipartFormData.CRLF)"))
+    encoded.append(try self.encode(string: "--\(self.boundary)"))
+    encoded.append(MultipartResponseParsing.CRLF)
     encoded.append(try self.encode(string: bodyPart.headers()))
     encoded.append(self.encode(inputStream: bodyPart.inputStream, length: bodyPart.contentLength))
-    encoded.append(try self.encode(string: "\(MultipartFormData.CRLF)"))
+    encoded.append(MultipartResponseParsing.CRLF)
 
     return encoded
   }
@@ -118,7 +123,7 @@ public final class MultipartFormData {
 
     var encoded = Data()
 
-    while (inputStream.hasBytesAvailable) {
+    while inputStream.hasBytesAvailable {
       var buffer = [UInt8](repeating: 0, count: 1024)
       let bytesRead = inputStream.read(&buffer, maxLength: 1024)
 
@@ -149,8 +154,7 @@ public final class MultipartFormData {
 
 extension MultipartFormData: Hashable {
   public static func == (lhs: MultipartFormData, rhs: MultipartFormData) -> Bool {
-    lhs.boundary == rhs.boundary &&
-    lhs.bodyParts == rhs.bodyParts
+    lhs.boundary == rhs.boundary && lhs.bodyParts == rhs.bodyParts
   }
 
   public func hash(into hasher: inout Hasher) {
@@ -162,18 +166,20 @@ extension MultipartFormData: Hashable {
 /// MARK: - BodyPart
 
 /// A structure representing a single part of multi-part form data.
-fileprivate struct BodyPart: Hashable {
+private struct BodyPart: Hashable {
   let name: String
   let inputStream: InputStream
   let contentLength: UInt64
   let contentType: String?
   let filename: String?
 
-  init(name: String,
-       inputStream: InputStream,
-       contentLength: UInt64,
-       contentType: String?,
-       filename: String?) {
+  init(
+    name: String,
+    inputStream: InputStream,
+    contentLength: UInt64,
+    contentType: String?,
+    filename: String?
+  ) {
     self.name = name
     self.inputStream = inputStream
     self.contentLength = contentLength
@@ -186,13 +192,13 @@ fileprivate struct BodyPart: Hashable {
     if let filename = self.filename {
       headers += "; filename=\"\(filename)\""
     }
-    headers += "\(MultipartFormData.CRLF)"
+    headers += "\r\n"
 
     if let contentType = self.contentType {
-      headers += "Content-Type: \(contentType)\(MultipartFormData.CRLF)"
+      headers += "Content-Type: \(contentType)\r\n"
     }
 
-    headers += "\(MultipartFormData.CRLF)"
+    headers += "\r\n"
 
     return headers
   }
