@@ -7,8 +7,10 @@
 /// # See Also
 /// [GraphQLSpec - Enums](https://spec.graphql.org/draft/#sec-Enums)
 public protocol EnumType:
+  Sendable,
   RawRepresentable,
   CaseIterable,
   JSONEncodable,
-  GraphQLOperationVariableValue
+  GraphQLOperationVariableValue,
+  GraphQLOperationVariableListElement
 where RawValue == String {}
